move logos assets to images folder
